refactor(): 코드 정리
