+++ src/main/java/lotto/domain/LottoGameResult.java
@@ -3,7 +3,6 @@ package lotto.domain;
 import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 public class LottoGameResult {
